Cleaning up the conversion logic so that the input-number logic stays in the input-number file.
